refactor(sucore): move Thor-specific ThorRootService into :app; add reusable RootService.enforceCaller()
